bug fix in g_code Visualiser and Started on  g code to point cloud
The basic bugs were fixed inside the you could visualise that result from movements in the Z axes which I was unaware off as well as removing the initial start and end lines of the print which resulted in the large lines of the sides. Basic work has been done into the G chord to point cloud program as initial inspection of the data was needed. Notable also been added to journal